requirements.txt
streamlit==1.29.0
anthropic==0.18.1
python-dotenv==1.0.0
httpx==0.25.2
Pillow==10.1.0
PyPDF2==3.0.1
reportlab==4.0.7
numpy==1.24.3
pandas==2.1.4
beautifulsoup4==4.12.2
python-dateutil==2.8.2
audio-recorder-streamlit==0.0.8
SpeechRecognition==3.10.0
gtts==2.4.0
python-docx==1.1.0
